202 support
All modes: legacy ticketless, legacy ticketed, and new

Currently it's the responsibility of the thing-that-calls StandardConsumer
to handle the actual re-request, since (especially in the case of a POST)
it is beyond StandardConsumer's scope.  There'll probably be some
librarification of that later to avoid duplicated code, but at this point
I don't know what code will actually end up being duplicated.